+++ src/main/java/me/tsukanov/counter/ui/SettingsActivity.java
@@ -9,7 +9,6 @@ import android.content.pm.PackageManager.NameNotFoundException;
 import android.content.res.Configuration;
 import android.os.Build;
 import android.os.Bundle;
-import android.preference.EditTextPreference;
 import android.preference.Preference;
 import android.preference.Preference.OnPreferenceChangeListener;
 import android.preference.Preference.OnPreferenceClickListener;
@@ -27,23 +26,30 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.util.Log;
 import android.widget.Toast;
 
 import me.tsukanov.counter.CounterApplication;
 import me.tsukanov.counter.R;
 
+
 public class SettingsActivity extends PreferenceActivity implements OnPreferenceChangeListener, SharedPreferences.OnSharedPreferenceChangeListener {
     private static final String KEY_REMOVE_COUNTERS = "removeCounters";
     private static final String KEY_VERSION = "version";
     private static final String KEY_THEME = "theme";
+    // ADDING CODE
+    private static final String KEY_COUNT_AMOUNT = "countAmount";
 
     private static final String THEME_DARK = "dark";
     private static final String THEME_LIGHT = "light";
 
+    private static final int MIN_COUNT_TICKS = 1;
+    private static final int MAX_COUNT_TICKS = 100;
+
+
     private SharedPreferences mSharedPref;
     private AppCompatDelegate mDelegate;
     private SettingsFragment fragment;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -85,9 +91,12 @@ public class SettingsActivity extends PreferenceActivity implements OnPreference
 
     private String getCurrentThemeName() {
         switch (mSharedPref.getString(KEY_THEME, THEME_LIGHT)) {
-            case THEME_LIGHT: return getResources().getString(R.string.settings_theme_light);
-            case THEME_DARK: return getResources().getString(R.string.settings_theme_dark);
-            default: return "Unknown";
+            case THEME_LIGHT:
+                return getResources().getString(R.string.settings_theme_light);
+            case THEME_DARK:
+                return getResources().getString(R.string.settings_theme_dark);
+            default:
+                return "Unknown";
         }
     }
 
@@ -143,16 +152,35 @@ public class SettingsActivity extends PreferenceActivity implements OnPreference
         return false;
     }
 
+
+//// ADDING CODE
+//editTextPreference.setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
+//        @Override
+//        public boolean onPreferenceChange(Preference preference, Object newValue) {
+//            int val = Integer.parseInt(newValue.toString());
+//            if ((val > minPort) && (val < maxPort)) {
+//
+//                Log.d(LOGTAG, "Value saved: " + val);
+//                return true;
+//            }
+//            else {
+//                // invalid you can show invalid message
+//                Toast.makeText(getApplicationContext(), "error text", Toast.LENGTH_LONG).show();
+//                return false;
+//            }
+//        }
+//    });
+
     public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
         if (key.equals(KEY_THEME)) {
             Preference pref;
             if (Build.VERSION.SDK_INT < Build.VERSION_CODES.HONEYCOMB) {
-               pref = findPreference(KEY_THEME);
+                pref = findPreference(KEY_THEME);
             } else {
                 pref = fragment.findPreference(KEY_THEME);
             }
             if (pref != null) {
-               pref.setSummary(getCurrentThemeName());
+                pref.setSummary(getCurrentThemeName());
             }
             // Restart to apply new theme, go back to this settings screen
             TaskStackBuilder.create(this)
@@ -272,6 +300,8 @@ public class SettingsActivity extends PreferenceActivity implements OnPreference
     @TargetApi(Build.VERSION_CODES.HONEYCOMB)
     public static class SettingsFragment extends PreferenceFragment {
         private OnPreferenceClickListener mOnRemoveCountersClickListener;
+        // ADDING CODE
+        private OnPreferenceClickListener mEditTextNumberChangedListener;
         private String mAppVersion;
         private String mTheme;
 
@@ -294,6 +324,49 @@ public class SettingsActivity extends PreferenceActivity implements OnPreference
             findPreference(KEY_VERSION).setSummary(mAppVersion);
             findPreference(KEY_THEME).setSummary(mTheme);
             findPreference(KEY_REMOVE_COUNTERS).setOnPreferenceClickListener(mOnRemoveCountersClickListener);
+            // ADDING CODE
+            findPreference(KEY_COUNT_AMOUNT).setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
+                @Override
+                public boolean onPreferenceChange(Preference preference, Object value) {
+
+                    int val = Integer.parseInt((String) value);
+                    if (val < MIN_COUNT_TICKS) {
+                        val = MIN_COUNT_TICKS;
+                    } else if (val > MAX_COUNT_TICKS) {
+                        val = MAX_COUNT_TICKS;
+                    }
+
+                    SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(getActivity());
+                    settings.edit().putString(KEY_COUNT_AMOUNT, "" + val).apply();
+                    return false;
+
+
+//                    if(newValue!=oldValue){
+//
+//                        settings.edit().putString(KEY_COUNT_AMOUNT,""+newValue).apply();
+//                        return false;
+//                    } else {
+//                        return true;
+//
+//                    }
+
+//                    int val2 = Integer.parseInt(settings.getString(KEY_COUNT_AMOUNT,"1"));
+
+//                    SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(getContext());
+//                    SharedPreferences.Editor prefEditor = sharedPref.edit(); // Get preference in editor mode
+//                    prefEditor.putString("your_edit_text_pref_key", "DEFAULT-VALUE"); // set your default value here (could be empty as well)
+//                    prefEditor.commit(); // finally save changes
+
+
+//                    try {
+//                        Toast.makeText(getContext(), "Value saved: " + newValue, Toast.LENGTH_LONG).show();
+//                    } catch (Exception e) {
+//
+//                    }
+//                    return false;
+
+                }
+            });
         }
     }
 }